moved all of the base plugin classes to their own modules
