+++ .flake8
@@ -10,5 +10,6 @@ rst-substitutions=version
 statistics=True
 per-file-ignores =
   /*/interfaces/__init__.py:F401
+  /*/interfaces/class_based/bases/__init__.py:F401
   tests/unit_tests/utils/types/test_datatree_ditto_FROM_XARRAY.py: D100,D101,D102,D103,D400,E501,E701
   tests/unit_tests/utils/types/test_datatree_ditto_mapping_FROM_XARRAY.py: D100,D101,D102,D103,D400,D202,D401
