+++ geoips/interfaces/class_based/bases/sector_spec_generators.py
@@ -1,0 +1,14 @@
+# # # This source code is subject to the license referenced at
+# # # https://github.com/NRLMMD-GEOIPS.
+
+"""Sector spec generators plugin class."""
+
+from geoips.interfaces.class_based_plugin import BaseClassPlugin
+
+
+class BaseSectorSpecGeneratorPlugin(BaseClassPlugin, abstract=True):
+    """Base class for GeoIPS sector_spec_generator plugins."""
+
+    data_tree = False
+
+    pass
