+++ geoips/interfaces/class_based/bases/title_formatters.py
@@ -1,0 +1,14 @@
+# # # This source code is subject to the license referenced at
+# # # https://github.com/NRLMMD-GEOIPS.
+
+"""Title formatters interface class."""
+
+from geoips.interfaces.class_based_plugin import BaseClassPlugin
+
+
+class BaseTitleFormatterPlugin(BaseClassPlugin, abstract=True):
+    """Base class for GeoIPS title_formatter plugins."""
+
+    data_tree = False
+
+    pass
