+++ geoips/interfaces/class_based/colormappers.py
@@ -3,33 +3,13 @@
 
 """Colormappers interface class."""
 
-import xarray as xr
-
-from geoips.interfaces.class_based_plugin import BaseClassPlugin
 from geoips.interfaces.base import BaseClassInterface
-from geoips.utils.types.datatree_ditto import DataTreeDitto
-
-
-class BaseColormapperPlugin(BaseClassPlugin, abstract=True):
-    """Base class for GeoIPS colormapper plugins."""
-
-    data_tree = False
-
-    def _post_call(self, data=None, *args, _obp_initiated=False, **kwargs):
-        """Wrap colormapper dict output into ``DataTreeDitto`` for OBP.
-
-        Colormappers return a ``dict`` with matplotlib colormap information.
-        OBP wraps this into ``DataTreeDitto`` with ``mpl_colors_info`` attrs
-        so downstream steps can extract it.  SSP receives the raw dict.
-        """
-        if _obp_initiated and isinstance(data, dict):
-            ds = xr.Dataset(attrs={"mpl_colors_info": data})
-            return DataTreeDitto(ds, name=getattr(self, "name", "colormapper"))
-        return data
 
 
 class ColormappersInterface(BaseClassInterface):
     """Interface for the colormapper to apply to the resulting product."""
+
+    from geoips.interfaces.class_based.bases import BaseColormapperPlugin
 
     name = "colormappers"
     plugin_class = BaseColormapperPlugin
