+++ geoips/interfaces/class_based/title_formatters.py
@@ -3,20 +3,13 @@
 
 """Title formatters interface class."""
 
-from geoips.interfaces.class_based_plugin import BaseClassPlugin
 from geoips.interfaces.base import BaseClassInterface
-
-
-class BaseTitleFormatterPlugin(BaseClassPlugin, abstract=True):
-    """Base class for GeoIPS title_formatter plugins."""
-
-    data_tree = False
-
-    pass
 
 
 class TitleFormattersInterface(BaseClassInterface):
     """Interface for creating GeoIPS formatted titles."""
+
+    from geoips.interfaces.class_based.bases import BaseTitleFormatterPlugin
 
     name = "title_formatters"
     plugin_class = BaseTitleFormatterPlugin
